requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Kafka
confluent-kafka==2.3.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Storage
minio==7.2.3

# Parsing libraries
python-docx==1.1.0
pdfplumber==0.10.3
PyMuPDF==1.23.21
python-pptx==0.6.23
openpyxl==3.1.2
pytesseract==0.3.10

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
structlog==24.1.0

# Metrics
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
